Changed all u32 types to usize to avoid type conversions.

+++ due/src/packets/dist_generator.rs
@@ -13,11 +13,11 @@ where
     A: Distribution<Time>,
     B: Distribution<f64>,
 {
-    element_id: u32,
+    element_id: usize,
     initial_delay: Time,
     arr_interval_dist: Box<dyn Fn() -> A>,
     packet_size_dist: Box<dyn Fn() -> B>,
-    packets_sent: u32,
+    packets_sent: usize,
     pub sender: UnboundedSender<Packet>,
     receiver: UnboundedReceiver<Packet>,
 }
@@ -42,7 +42,7 @@ where
     B: Distribution<f64>,
 {
     pub fn new(
-        element_id: u32,
+        element_id: usize,
         initial_delay: Time,
         arr_interval_dist: Box<dyn Fn() -> A>,
         packet_size_dist: Box<dyn Fn() -> B>,
@@ -83,7 +83,7 @@ where
             let interval = (self.arr_interval_dist)().sample(&mut *sim.shared().rng.borrow_mut());
             sim.advance(interval).await;
             let packet_size =
-                (self.packet_size_dist)().sample(&mut *sim.shared().rng.borrow_mut()) as u32;
+                (self.packet_size_dist)().sample(&mut *sim.shared().rng.borrow_mut()) as usize;
 
             let mut packet = Packet::new(
                 packet_size,
