+++ due/src/switches/simple_packet_switch.rs
@@ -7,13 +7,13 @@ use crate::{sim::SimContext, Element};
 use tokio::sync::mpsc::{unbounded_channel, UnboundedReceiver, UnboundedSender};
 
 pub struct SimplePacketSwitch {
-    element_id: u32,
+    element_id: usize,
     /// the number of packets received by the switch
-    packets_received: u32,
+    packets_received: usize,
     /// the fib demux of the switch,
-    pub fib: Vec<u32>,
+    pub fib: Vec<usize>,
     /// a closure that maps a flow_id to a class_id
-    pub flow_classes: Box<dyn Fn(u32) -> u32>,
+    pub flow_classes: Box<dyn Fn(usize) -> usize>,
     /// the output ports of the switch, with consecutive port ids start from 0
     pub ports: Vec<Port>,
     /// senders for sending outbound packets to ports or schedulers
@@ -41,10 +41,10 @@ impl Element for SimplePacketSwitch {
 
 impl SimplePacketSwitch {
     pub fn new(
-        element_id: u32,
-        nports: u32,
+        element_id: usize,
+        nports: usize,
         port_rate: f64,
-        buffer_size: u32,
+        buffer_size: usize,
     ) -> SimplePacketSwitch {
         let mut ports = Vec::new();
         let mut senders = Vec::new();
@@ -77,8 +77,8 @@ impl SimplePacketSwitch {
                 let flow_class = (self.flow_classes)(packet.flow_id);
 
                 // forwards packets to their corresponding ports
-                if let Some(&port_id) = self.fib.get(flow_class as usize) {
-                    let _ = self.senders.get_mut(port_id as usize).unwrap().send(packet);
+                if let Some(&port_id) = self.fib.get(flow_class) {
+                    let _ = self.senders.get_mut(port_id).unwrap().send(packet);
                 } else {
                     println!("Wrong fib demux in SimplePacketSwitch {}.", self.element_id);
                 }
